Checkando se a senha e usuário estão corretos

+++ Trabalho_Pesquisa/CriarConta.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Text;
+using System.Threading;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 
@@ -14,6 +15,7 @@ namespace Trabalho_Pesquisa
 {
     public partial class CriarConta : Form
     {
+        Thread nt;
         public CriarConta()
         {
             InitializeComponent();
@@ -42,11 +44,34 @@ namespace Trabalho_Pesquisa
             byte[] hash = sha512.ComputeHash(bytes);
             string senhaCript = Convert.ToBase64String(hash);
 
-            //Arquivo Conta
-            StreamWriter sw = new StreamWriter($"C:\\Users\\Alunos\\Documents\\Saves NewSearch\\{username}Conta.txt");
-            sw.WriteLine(senhaCript);
-            sw.Close();
-            MessageBox.Show("Nova Conta Criada");
+            if (!File.Exists($"C:\\Users\\Alunos\\Documents\\Saves NewSearch\\{username}Conta.txt"))
+            {
+                //Arquivo Conta
+                StreamWriter sw = new StreamWriter($"C:\\Users\\Alunos\\Documents\\Saves NewSearch\\{username}Conta.txt");
+                sw.WriteLine(senhaCript);
+                sw.Close();
+                MessageBox.Show("Nova Conta Criada");
+            } else
+            {
+                MessageBox.Show("Conta já existente");
+            }
+        }
+
+        private void txtUsername_TextChanged(object sender, EventArgs e)
+        {
+
+        }
+
+        private void btnjanelaSignUp_Click(object sender, EventArgs e)
+        {
+            this.Close();
+            nt = new Thread(Login);
+            nt.SetApartmentState(ApartmentState.STA);
+            nt.Start();
+        }
+        private void Login()
+        {
+            Application.Run(new Login());
         }
     }
 }
